requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
anthropic>=0.40.0
pydantic-settings>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
respx>=0.22.0
